+++ beaker/progress.py
@@ -22,7 +22,7 @@ from rich.table import Table
 from rich.text import Text
 
 
-class DownloadUploadColumn(DownloadColumn):
+class ImageDownloadUploadColumn(DownloadColumn):
     def render(self, task: Task) -> Text:
         if task.total is None or int(task.total) == 1:
             return Text("")
@@ -128,7 +128,7 @@ def get_dataset_sync_progress(quiet: bool = False) -> Progress:
         "[progress.percentage]{task.percentage:>3.0f}%",
         TimeElapsedColumn(),
         TimeRemainingColumn(),
-        DownloadUploadColumn(),
+        DownloadColumn(),
         disable=quiet,
     )
 
@@ -140,7 +140,7 @@ def get_sized_dataset_fetch_progress(quiet: bool = False) -> Progress:
         "[progress.percentage]{task.percentage:>3.0f}%",
         TimeElapsedColumn(),
         TimeRemainingColumn(),
-        DownloadUploadColumn(),
+        DownloadColumn(),
         disable=quiet,
     )
 
@@ -150,7 +150,7 @@ def get_unsized_dataset_fetch_progress(quiet: bool = False) -> Progress:
         "[progress.description]{task.description}",
         SpinnerColumn(),
         TimeElapsedColumn(),
-        DownloadUploadColumn(),
+        FileSizeColumn(),
         disable=quiet,
     )
 
@@ -161,7 +161,7 @@ def get_image_upload_progress(quiet: bool = False) -> Progress:
         BarColumn(),
         "[progress.percentage]{task.percentage:>3.0f}%",
         TimeRemainingColumn(),
-        DownloadUploadColumn(),
+        ImageDownloadUploadColumn(),
         disable=quiet,
     )